backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.2
DeepFace==0.0.75
tensorflow==2.13.0
werkzeug==2.3.7
python-dotenv==1.0.0
stripe==5.15.0
python-dateutil==2.8.2

